Fixed "WHY?!" bug, caused by python-sql Unicode issues
SQL text with invalid UTF-8 can't be passed directly to a custom
functions, only 'None' will be received. The UTF-8 needs to be converted
to a blob first and then back to text once the custom function returned.
this can be done with:

cast(py_dirname(cast(column_name AS BLOB)) AS TEXT)